Rename analyzer_lsp -> analyzer
