+++ src/extension/analysis/analyzer.ts
